+++ src/Ninject.Test/Unit/PropertyInjectionStrategyTests.cs
@@ -150,7 +150,7 @@ namespace Ninject.Tests.Unit.PropertyInjectionStrategyTests
         protected Mock<ITarget> barTargetMock;
         protected InstanceReference reference;
         protected FakePropertyInjectionDirective[] directives;
-        protected IEnumerable<IParameter> parameters;
+        protected IReadOnlyList<IParameter> parameters;
 
         public WhenActivateIsCalled_WithProjectInjectionDirectivesAndWithPropertyValuesFullMatch()
         {
@@ -241,7 +241,7 @@ namespace Ninject.Tests.Unit.PropertyInjectionStrategyTests
         protected Mock<ITarget> gooTargetMock;
         protected InstanceReference reference;
         protected FakePropertyInjectionDirective[] directives;
-        protected IEnumerable<IParameter> parameters;
+        protected IReadOnlyList<IParameter> parameters;
 
         public WhenActivateIsCalled_WithProjectInjectionDirectivesAndWithPropertyValuesPartialMatch()
         {
@@ -348,7 +348,7 @@ namespace Ninject.Tests.Unit.PropertyInjectionStrategyTests
 
         public WhenActivateIsCalled_WithoutProjectInjectionDirectivesAndWithoutPropertyValues()
         {
-            IEnumerable<IParameter> parameters = new List<IParameter>
+            IReadOnlyList<IParameter> parameters = new List<IParameter>
                 {
                     new ConstructorArgument("A", "B")
                 };
@@ -388,7 +388,7 @@ namespace Ninject.Tests.Unit.PropertyInjectionStrategyTests
         protected Mock<PropertyInjector> barInjectorMock;
         protected InstanceReference reference;
         protected PropertyInjectionDirective[] directives;
-        protected IEnumerable<IParameter> parameters;
+        protected IReadOnlyList<IParameter> parameters;
 
         public WhenActivateIsCalled_WithoutProjectInjectionDirectivesAndWithPropertyValues()
         {
@@ -453,7 +453,7 @@ namespace Ninject.Tests.Unit.PropertyInjectionStrategyTests
         protected Dummy instance;
         protected InstanceReference reference;
         private string exceptionMessage;
-        protected IEnumerable<IParameter> parameters;
+        protected IReadOnlyList<IParameter> parameters;
 
         public WhenActivateIsCalled_PropertyValueCannotBeResolvedToProperty()
         {